requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymupdf
pytesseract
Pillow
transformers
sentencepiece
torch
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
python-docx
fpdf2
orjson
regex
